+++ octodns/zone/validator/__init__.py
@@ -1,0 +1,14 @@
+#
+#
+#
+
+from .base import ValidationReason, ZoneValidator, ZoneValidatorRegistry
+
+__all__ = [
+    'ValidationReason',
+    'ZoneValidator',
+    'ZoneValidatorRegistry',
+    'zone_validators',
+]
+
+zone_validators = ZoneValidatorRegistry()
